Add files via upload
The bitwise AND operator & then performs a logical AND operation between the shifted data and the mask 0x01. The mask 0x01 has all bits set to zero, except for the least significant bit, which is set to one. This means that the result of the AND operation will be either zero or one, depending on the value of the i-th bit of data.